backend-alessandra/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
mysql-connector-python
